More work on interaction between server and client
I set it up so that the server returns both a message and details (for
the response), when you do a delete.

I added a DOB, Home Phone, Number of Children, and Hair Color to the
array returned by getAll.php (the array is just for demo purposes).

I introduced use of some new HTML5 input types.

I fiddled with the SWAL dialog for record deletion.  It still needs
work but it is at least “demoable”.